# Conflicts:
#	matlab_runners/Feature Point Experiments/results/landmark_detections.txt

+++ lib/local/LandmarkDetector/include/Patch_experts.h
@@ -98,7 +98,7 @@ public:
 	// Additionally returns the transform from the image coordinates to the response coordinates (and vice versa).
 	// The computation also requires the current landmark locations to compute response around, the PDM corresponding to the desired model, and the parameters describing its instance
 	// Also need to provide the size of the area of interest and the desired scale of analysis
-	void Response(vector<cv::Mat_<float> >& patch_expert_responses, cv::Matx22f& sim_ref_to_img, cv::Matx22f& sim_img_to_ref, const cv::Mat_<uchar>& grayscale_image, 
+	void Response(vector<cv::Mat_<float> >& patch_expert_responses, cv::Matx22f& sim_ref_to_img, cv::Matx22f& sim_img_to_ref, const cv::Mat_<float>& grayscale_image, 
 							 const PDM& pdm, const cv::Vec6f& params_global, const cv::Mat_<float>& params_local, int window_size, int scale);
 
 	// Getting the best view associated with the current orientation
